python-modbus-service/requirements.txt
pymodbus==3.5.4
requests==2.31.0
APScheduler==3.10.4
python-dotenv==1.0.0 